Moves Crypt4GH utilities to shared module
Extracts Crypt4GH plaintext cleanup functions and constants from the
remote execution module into the centralized utility module. This refactoring
reduces duplication, improves code maintainability, and establishes a single
source of truth for Crypt4GH-related utilities.
